+++ src/neuroconv/datainterfaces/ophys/caiman/caimandatainterface.py
@@ -1,13 +1,9 @@
-from roiextractors import CaimanSegmentationExtractor
-
 from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
 from ....utils import FilePathType
 
 
 class CaimanSegmentationInterface(BaseSegmentationExtractorInterface):
     """Data interface for CaimanSegmentationExtractor."""
-
-    SegX = CaimanSegmentationExtractor
 
     def __init__(self, file_path: FilePathType, verbose: bool = True):
         super().__init__(file_path=file_path)