+++ tests/test_internals/test_interfaces.py
@@ -28,8 +28,8 @@ class TestAssertions(TestCase):
     def test_import_assertions(self):
         if platform == "darwin" and version.parse(python_version()) < version.parse("3.8"):
             with self.assertRaisesWith(
-                exc_type=AssertionError,
-                exc_msg="The sonpy package (CED dependency) is not available on Mac for Python versions below 3.8!",
+                exc_type=ModuleNotFoundError,
+                exc_msg=("\nThe package 'sonpy' is not available on the darwin platform for Python version 3.7!"),
             ):
                 CEDRecordingInterface.get_all_channels_info(file_path="does_not_matter.smrx")
         else: